fix: fix/go-trivy-build
