+++ go/main.go
@@ -257,137 +257,3 @@ func (m *Go) RunWorkflowEntryStage(
 	// Return the stats file
 	return statsFile, nil
 }
-
-// Lint runs the linter on the provided source code
-func (m *Go) Lint(
-	ctx context.Context,
-	src *dagger.Directory,
-	// +optional
-	// +default="500s"
-	timeout string,
-) *dagger.Container {
-
-	golangciLintRunOpts := dagger.GolangciLintRunOpts{
-		Timeout: timeout,
-	}
-
-	return dag.GolangciLint().Run(src, golangciLintRunOpts)
-}
-
-func (m *Go) KoBuildAndScan(
-	ctx context.Context,
-	src *dagger.Directory,
-	// +optional
-	// +default="GITHUB_TOKEN"
-	tokenName string,
-	token *dagger.Secret,
-	// +optional
-	// +default="ko.local"
-	repo string,
-	// +optional
-	// +default="."
-	buildArg string,
-	// +optional
-	// +default="v0.17.1"
-	koVersion string,
-	// +optional
-	// +default="true"
-	push string,
-	// +optional
-	// +default="HIGH,CRITICAL"
-	severityFilter string, // Comma-separated list of severities to filter (e.g., "HIGH,CRITICAL")
-) (string, error) {
-	// Step 1: Build the image using KoBuild and push it to the remote registry
-	imageAddress, err := m.KoBuild(ctx, src, tokenName, token, repo, buildArg, koVersion, push)
-	if err != nil {
-		return "", fmt.Errorf("error building and pushing image: %w", err)
-	}
-
-	// Step 2: Scan the remote image using Trivy
-	scanResult, err := m.ScanRemoteImage(ctx, imageAddress, severityFilter)
-	if err != nil {
-		return "", fmt.Errorf("error scanning remote image: %w", err)
-	}
-
-	// Step 3: Parse the Trivy scan report and search for vulnerabilities
-	vulnerabilities, err := m.SearchVulnerabilities(ctx, scanResult, severityFilter)
-	if err != nil {
-		return "", fmt.Errorf("error searching vulnerabilities: %w", err)
-	}
-
-	// Step 4: Return the vulnerabilities found
-	if len(vulnerabilities) > 0 {
-		return fmt.Sprintf("Found vulnerabilities: %v", vulnerabilities), nil
-	}
-
-	return "No vulnerabilities found.", nil
-}
-
-func (m *Go) Test(
-	ctx context.Context,
-	src *dagger.Directory,
-	goVersion string, // Go version to use for testing
-	// +optional
-	// +default="./..."
-	testArg string, // Arguments for `go test`
-) (string, error) {
-	// Create a container with the specified Go version
-	container := dag.Container().
-		From(fmt.Sprintf("golang:%s", goVersion)). // Use the specified Go version
-		WithDirectory("/src", src).
-		WithWorkdir("/src")
-
-	// Run Go tests with coverage
-	output, err := container.
-		WithExec([]string{"go", "test", "-cover", testArg}).
-		Stdout(ctx)
-	if err != nil {
-		return "", fmt.Errorf("error running tests: %w", err)
-	}
-
-	return output, nil
-}
-
-func (m *Go) RunWorkflowContainerStage(
-	ctx context.Context,
-	src *dagger.Directory,
-	// +optional
-	// +default="GITHUB_TOKEN"
-	tokenName string,
-	token *dagger.Secret,
-	// +optional
-	// +default="ko.local"
-	repo string,
-	// +optional
-	// +default="."
-	buildArg string,
-	// +optional
-	// +default="v0.17.1"
-	koVersion string,
-	// +optional
-	// +default="true"
-	push string,
-	// +optional
-	// +default="HIGH,CRITICAL"
-	severityFilter string, // Comma-separated list of severities to filter (e.g., "HIGH,CRITICAL")
-) (string, error) {
-	// Step 1: Build the image with ko and scan it for vulnerabilities
-	scanResult, err := m.KoBuildAndScan(ctx, src, tokenName, token, repo, buildArg, koVersion, push, severityFilter)
-	if err != nil {
-		return "", fmt.Errorf("error during KoBuildAndScan: %w", err)
-	}
-
-	// Step 2: Check if vulnerabilities were found
-	if strings.Contains(scanResult, "Found vulnerabilities:") {
-		return scanResult, nil // Return the vulnerabilities found
-	}
-
-	// Step 3: If no vulnerabilities were found, build and push the image to the remote registry
-	imageAddress, err := m.KoBuild(ctx, src, tokenName, token, repo, buildArg, koVersion, push)
-	if err != nil {
-		return "", fmt.Errorf("error building and pushing image: %w", err)
-	}
-
-	// Step 4: Return the image address
-	return fmt.Sprintf("Image built and pushed successfully: %s", imageAddress), nil
-}
