+++ go/verify.go
@@ -1,0 +1,48 @@
+package main
+
+import (
+	"context"
+	"dagger/go/internal/dagger"
+	"fmt"
+)
+
+func (m *Go) Test(
+	ctx context.Context,
+	src *dagger.Directory,
+	goVersion string, // Go version to use for testing
+	// +optional
+	// +default="./..."
+	testArg string, // Arguments for `go test`
+) (string, error) {
+	// Create a container with the specified Go version
+	container := dag.Container().
+		From(fmt.Sprintf("golang:%s", goVersion)). // Use the specified Go version
+		WithDirectory("/src", src).
+		WithWorkdir("/src")
+
+	// Run Go tests with coverage
+	output, err := container.
+		WithExec([]string{"go", "test", "-cover", testArg}).
+		Stdout(ctx)
+	if err != nil {
+		return "", fmt.Errorf("error running tests: %w", err)
+	}
+
+	return output, nil
+}
+
+// Lint runs the linter on the provided source code
+func (m *Go) Lint(
+	ctx context.Context,
+	src *dagger.Directory,
+	// +optional
+	// +default="500s"
+	timeout string,
+) *dagger.Container {
+
+	golangciLintRunOpts := dagger.GolangciLintRunOpts{
+		Timeout: timeout,
+	}
+
+	return dag.GolangciLint().Run(src, golangciLintRunOpts)
+}
